+++ sampledb/logic/schemas/validate_schema.py
@@ -6,6 +6,7 @@ Implementation of validate_schema(schema)
 
 import datetime
 import typing
+import urllib.parse
 import re
 
 from ..errors import ValidationError
@@ -162,6 +163,7 @@ def _validate_object_schema(schema: dict, path: typing.List[str]) -> None:
         valid_keys.add('displayProperties')
         valid_keys.add('batch')
         valid_keys.add('batch_name_format')
+        valid_keys.add('notebookTemplates')
     required_keys = {'type', 'title', 'properties'}
     schema_keys = set(schema.keys())
     invalid_keys = schema_keys - valid_keys
@@ -230,6 +232,9 @@ def _validate_object_schema(schema: dict, path: typing.List[str]) -> None:
             schema['batch_name_format'].format(1)
         except (ValueError, KeyError):
             raise ValidationError('invalid batch_name_format', path)
+
+    if 'notebookTemplates' in schema:
+        _validate_notebook_templates(schema['notebookTemplates'])
 
 
 def _validate_text_schema(schema: dict, path: typing.List[str]) -> None:
@@ -379,3 +384,53 @@ def _validate_sample_schema(schema: dict, path: typing.List[str]) -> None:
         raise ValidationError('missing keys in schema: {}'.format(missing_keys), path)
     if 'note' in schema and not isinstance(schema['note'], str):
         raise ValidationError('note must be str', path)
+
+
+def _validate_notebook_templates(notebook_templates: typing.Any) -> None:
+    """
+    Validate the given notebook templates and raise a ValidationError if they are invalid.
+
+    :param notebook_templates: the sampledb object schema
+    :raise ValidationError: if the notebook templates are invalid
+    """
+    if not isinstance(notebook_templates, list):
+        raise ValidationError('notebookTemplates must be a list', ['notebookTemplates'])
+
+    for notebook_index, notebook_template in enumerate(notebook_templates):
+        path = ['notebookTemplates', str(notebook_index)]
+        if not isinstance(notebook_template, dict):
+            raise ValidationError('notebook template must be a dict', path)
+        valid_keys = {'title', 'url', 'params'}
+        required_keys = valid_keys
+        schema_keys = set(notebook_template.keys())
+        invalid_keys = schema_keys - valid_keys
+        if invalid_keys:
+            raise ValidationError('unexpected keys in notebook template: {}'.format(invalid_keys), path)
+        missing_keys = required_keys - schema_keys
+        if missing_keys:
+            raise ValidationError('missing keys in notebook template: {}'.format(missing_keys), path)
+        if not isinstance(notebook_template['title'], str):
+            raise ValidationError('notebook template title must be str', path)
+        if not isinstance(notebook_template['url'], str):
+            raise ValidationError('notebook template url must be str', path)
+        if not notebook_template['url'].endswith('.ipynb'):
+            raise ValidationError('notebook template url must end with .ipynb', path)
+        base_url = 'https://iffjupyter.fz-juelich.de/templates/t/'
+        test_url = urllib.parse.urljoin(base_url, notebook_template['url'])
+        if not test_url.startswith(base_url) or urllib.parse.urlparse(base_url).netloc != urllib.parse.urlparse(test_url).netloc:
+            raise ValidationError('notebook template url must be relative', path)
+        if not isinstance(notebook_template['params'], dict):
+            raise ValidationError('notebook template params must be a dict', path)
+        for param_name, param_value in notebook_template['params'].items():
+            if not isinstance(param_name, str):
+                raise ValidationError('notebook template param names must be str', path)
+            if not isinstance(param_value, str) and not isinstance(param_value, list):
+                raise ValidationError('notebook template param values must be str or a list', path)
+            if isinstance(param_value, list):
+                for param_step in param_value:
+                    if not isinstance(param_step, str) and not isinstance(param_step, int):
+                        raise ValidationError('notebook template param value steps must be str or int', path)
+            if isinstance(param_value, str):
+                valid_param_values = {'object_id'}
+                if param_value not in valid_param_values:
+                    raise ValidationError('notebook template param value must be a list or one of {}'.format(valid_param_values), path)